- Refactoring, renaming, moving and commenting as a result of Review - Added unit tests
