+++ CDP4RequirementsVerification/CDP4RequirementsVerification.csproj
@@ -1,0 +1,25 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net45;net451;net452;net46;net461;net462;net47;net471;net472;netstandard2.0</TargetFrameworks>
+    <Company>RHEA System S.A.</Company>
+    <Title>CDP4RequirementsVerification Community Edition</Title>
+    <VersionPrefix>1.0.0</VersionPrefix>
+    <Description>CDP4 Class Library that provides requirement verification</Description>
+    <Copyright>Copyright © RHEA System S.A.</Copyright>
+    <Authors>Sam, Alex, Alexander, Yevhen</Authors>
+    <PackageId>CDP4RequirementsVerification-CE</PackageId>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <PackageProjectUrl>http://sdk.cdp4.org</PackageProjectUrl>
+    <PackageIconUrl>https://raw.githubusercontent.com/RHEAGROUP/CDP4-SDK-Community-Edition/master/docs/cdp4-icon.png</PackageIconUrl>
+    <RepositoryUrl>https://github.com/RHEAGROUP/CDP4-SDK-Community-Edition.git</RepositoryUrl>
+    <PackageTags>CDP CDP4 ECSS-E-TM-10-25</PackageTags>
+    <PackageLicenseUrl>https://github.com/RHEAGROUP/CDP4-SDK-Community-Edition/blob/master/COPYING.LESSER</PackageLicenseUrl>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\CDP4Common\CDP4Common.csproj" />
+    <ProjectReference Include="..\CDP4Dal\CDP4Dal.csproj" />
+  </ItemGroup>
+
+</Project>
